Size is not grandeur, and territory does not make a nation edit

+++ 0x00-hello_world/6-size.c
@@ -7,10 +7,12 @@
 
 int main(void)
 {
-	printf("char size: %lu bytes\n", sizeof(char));
-	printf("int size: %lu bytes\n", sizeof(int));
-	printf("long size: %lu bytes\n", sizeof(long));
-	printf("long long size: %lu bytes\n", sizeof(long long));
-	printf("float size: %lu bytes\n", sizeof(float));
+
+	printf("char size: %lu bytes\n", (unsigned long)sizeof(char));
+	printf("int size: %lu bytes\n", (unsigned long)sizeof(int));
+	printf("long size: %lu bytes\n", (unsigned long)sizeof(long));
+	printf("long long size: %lu bytes\n", (unsigned long)sizeof(long long));
+	printf("float size: %lu bytes\n", (unsigned long)sizeof(float));
 	return (0);
+
 }
